requirements.txt
playwright==1.44.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
Pillow==10.3.0
imagehash==4.3.1
groq>=0.9.0
datasets==2.19.0
huggingface_hub==0.23.0
torch==2.2.2
open-clip-torch==2.24.0
scikit-image==0.23.2
opencv-python-headless==4.9.0.80
numpy==1.26.4
tqdm==4.66.4
fake-useragent==1.5.1
pyarrow==15.0.2
brotli==1.1.0
python-dotenv>=1.0.0
